Conflicts:
	java_ca/src/main/java/edu/nus/java_ca/controller/AdminLeaveController.java

+++ java_ca/src/main/java/edu/nus/java_ca/controller/AdminLeaveTypeController.java
@@ -1,8 +1,8 @@
 package edu.nus.java_ca.controller;
 
-import java.util.ArrayList;
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,12 +13,13 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.support.SessionStatus;
 import org.springframework.web.servlet.ModelAndView;
 
 import edu.nus.java_ca.model.LeaveBalance;
-import edu.nus.java_ca.model.Position;
 import edu.nus.java_ca.model.User;
 import edu.nus.java_ca.service.LeaveBalanceService;
+import edu.nus.java_ca.service.SessionManagement;
 import edu.nus.java_ca.service.UserService;
 
 @Controller
@@ -30,24 +31,30 @@ public class AdminLeaveTypeController {
 	
 	@Autowired
 	LeaveBalanceService lbService;
+	
+	@Autowired
+	SessionManagement sess;
 
 
 	@RequestMapping({ "/", "" })
-	public String dashboard(Model model) {
+	public String dashboard(Model model, HttpSession ses, SessionStatus status) {
+		if (!sess.isLoggedIn(ses, status)) return "redirect:/";
 		List<String> leavetypes = lbService.findAllLeaveTypes();
 		model.addAttribute("ltlist", leavetypes);
 		return "admin/leavetypes";
 	}
 	
 	@RequestMapping(value = "/add", method = RequestMethod.GET)
-	public ModelAndView addLeaveType() {
+	public ModelAndView addLeaveType(HttpSession ses, SessionStatus status) {
+		if (!sess.isLoggedIn(ses, status)) return new ModelAndView("redirect:/");
 		ModelAndView mav = new ModelAndView("admin/leave-type-form", "newlt", new LeaveBalance());
 		return mav;
 	}
 	
 	@RequestMapping(value = "/add", method = RequestMethod.POST)
 	public String saveMember(@ModelAttribute("newlt") @Valid LeaveBalance lb,
-			BindingResult bindingResult) {
+			BindingResult bindingResult, HttpSession ses, SessionStatus status) {
+		if (!sess.isLoggedIn(ses, status)) return "redirect:/";
 		if (bindingResult.hasErrors()) {
 			return "admin/leave-type-form";
 		}
@@ -63,8 +70,8 @@ public class AdminLeaveTypeController {
 	}
 	
 	@RequestMapping(value = "/delete/{eleave}")
-	public String deleteLeaveType(@PathVariable("eleave") String eleave) {
-		
+	public String deleteLeaveType(@PathVariable("eleave") String eleave, HttpSession ses, SessionStatus status) {
+		if (!sess.isLoggedIn(ses, status)) return "redirect:/";
 		List<User> userlist = uService.findAllWithDeleted();
 		for (User staffUser : userlist) {
 			lbService.deleteLeaveBalanceByType(eleave, staffUser);
